chore(forge): make remappings standalone command
